source/requirements.txt
speedtest-cli
instabot